Moved settings service to service folder
