+++ webapp/src/app/components/dialogs/settings/settings.component.ts
@@ -2,11 +2,11 @@ import { Component, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
 
-import { SettingsService } from '../../../directives/settings.service';
 import { BrowserService } from '../../../services/browser.service';
 import { ElectronService } from '../../../services/electron.service';
 import { Globals } from '../../../services/globals';
 import { HttpClientService } from '../../../services/http-client.service';
+import { SettingsService } from '../../../services/settings.service';
 import { SharedService } from '../../../services/shared-service';
 import { OverlayRefControl } from '../overlay/overlay-ref-control';
