+++ webapp/src/app/components/widgets/event-widget/event-editor/editor/event-editor.component.ts
@@ -2,8 +2,8 @@ import { CdkDrag, CdkDragDrop, CdkDropList } from '@angular/cdk/drag-drop';
 import { FlatTreeControl } from '@angular/cdk/tree';
 import { ChangeDetectionStrategy, Component, ElementRef, EventEmitter, Input, OnChanges, OnInit, Output, ViewChild } from '@angular/core';
 import { MatTreeFlatDataSource, MatTreeFlattener } from '@angular/material/tree';
-import { SettingsService } from '../../../../../directives/settings.service';
 import { destructureEventArray, EventArray } from '../../../../../models/events';
+import { SettingsService } from '../../../../../services/settings.service';
 import { SplitPaneComponent } from '../../../../split-pane/split-pane.component';
 import { AbstractEvent, EventType } from '../../event-registry/abstract-event';
 import { AddEventService } from '../add/add-event.service';
